BioDecoderForNC
Due to the large size of the library and input/output files, they cannot be uploaded to GitHub at this stage. Researchers interested in accessing the full BioDecoder Library and related data files are encouraged to contact the corresponding author. The resources will be made publicly available on a suitable hosting platform as soon as possible..